requirements.txt
flask
python-dotenv
azure-core
azure-ai-documentintelligence